feat: 온보딩 컴포넌트 분리
