+++ src/pages/Onboarding/UserRegistration/Profile/page.tsx
@@ -3,8 +3,9 @@ import { Progress } from "@/components/ui/progress";
 import { useNavigate } from "react-router-dom";
 import { ReactComponent as BackArrow } from "@/assets/svgs/BackArrow.svg";
 import { ConfirmButton } from "@/components/common/Button";
-import { Profile } from "@/components/common/Profile";
 import { useState } from "react";
+import { Gender } from "@/components/common/Gender";
+import { Age } from "@/components/common/Age";
 
 export const SelectProfilePage = () => {
   const navigate = useNavigate();
@@ -38,13 +39,19 @@ export const SelectProfilePage = () => {
             <span className="body1-medium text-grayscale-60">
               성별과 연령대를 알려주세요.
             </span>
+
             <div className="slide-up">
-              <Profile
-                selectedGender={selectedGender}
-                selectedAge={selectedAge}
-                onGenderSelect={setSelectedGender}
-                onAgeSelect={setSelectedAge}
-              />
+              <div className="flex flex-col gap-[0.87rem] mt-[2.12rem] mb-[2.88rem]">
+                <span className="body1-medium text-grayscale-80">성별</span>
+                <Gender
+                  selectedGender={selectedGender}
+                  onGenderSelect={setSelectedGender}
+                />
+              </div>
+              <div className="flex flex-col gap-[0.87em] mb-[56px]">
+                <span className="body1-medium text-grayscale-80">연령대</span>
+                <Age selectedAge={selectedAge} onAgeSelect={setSelectedAge} />
+              </div>
             </div>
           </div>
           <ConfirmButton
